Refactored dropdowns
